Allow multiple proxies to be deploy. Only keep proxy amount of proxy entries in the scrubber config
