+++ ovs/lib/tests/generic_tests/test_generic.py
@@ -18,29 +18,18 @@
 Generic test module
 """
 import os
-import re
 import time
 import datetime
 import unittest
-from threading import Event
-from ovs.dal.hybrids.diskpartition import DiskPartition
 from ovs.dal.hybrids.servicetype import ServiceType
-from ovs.dal.hybrids.vdisk import VDisk
 from ovs.dal.lists.servicetypelist import ServiceTypeList
-from ovs.dal.lists.storagerouterlist import StorageRouterList
 from ovs.dal.tests.helpers import DalHelper
 from ovs.extensions.db.arakooninstaller import ArakoonClusterConfig, ArakoonInstaller
 from ovs.extensions.generic.configuration import Configuration
 from ovs.extensions.generic.sshclient import SSHClient, UnableToConnectException
 from ovs_extensions.generic.tests.sshclient_mock import MockedSSHClient
-from ovs.extensions.generic.system import System
-from ovs_extensions.generic.threadhelpers import Waiter
-from ovs_extensions.services.mockups.systemd import SystemdMock
-from ovs.extensions.services.servicefactory import ServiceFactory
-from ovs.extensions.storageserver.tests.mockups import LockedClient
 from ovs.lib.generic import GenericController
 from ovs.lib.helpers.toolbox import Toolbox
-from ovs.lib.helpers.generic.scrubber import ScrubShared, StackWorker
 from ovs.lib.vdisk import VDiskController
 from ovs.log.log_handler import LogHandler
 
@@ -267,703 +256,6 @@ class Generic(unittest.TestCase):
                                                     metadata={'label': 'ss_c_{0}:30'.format(str(h)),
                                                               'is_consistent': True,
                                                               'timestamp': str(ts)})
-
-    def test_scrubbing(self):
-        """
-        Validates the scrubbing workflow
-        * Scenario 1: Validate disabled scrub task and single vDisk scrub logic
-        * Scenario 2: 1 vPool, 10 vDisks, 1 scrub role
-                      Scrubbing fails for 5 vDisks, check if scrubbing completed for all other vDisks
-                      Run scrubbing a 2nd time and verify scrubbing now works for failed vDisks
-        * Scenario 3: 1 vPool, 11 vDisks, 5 scrub roles (4 StorageRouters, one of then has 2 scrub roles)
-                      Check template vDisk is NOT scrubbed
-                      Check if vDisks are divided among all threads
-        * Scenario 4: 3 vPools, 15 vDisks, 5 scrub roles
-                      Validate 12 threads will be spawned and used out of a potential of 30 (5 scrub roles * 3 vPools * 2 threads per StorageRouter)
-                      We limit max amount of threads spawned per vPool to 2 in case 3 to 5 vPools are present
-        * Scenario 5: 2 vPools, 8 vDisks, 2 scrub roles
-                      Validate correct vDisks are scrubbed on expected location when specifying vpool_guids and/or vdisk_guids
-        * Scenario 6: Configure amount of threads per StorageRouter to 5 for SR1
-                      2 vPools, 20 vDisks, 2 scrub roles
-                      Validate expected amount of threads is spawned
-        * Scenario 7: 2 vPools, 20 vDisks, 2 scrub roles
-                      Validate special conditions that can occur
-                      - Proxy deployment fail (No thread should should and all should fail)
-                      - Proxy removal fail after scrub (All threads will try to clean up the proxy)
-        * Scenario 8: 2 vPools, 20 vDisks, 2 scrub roles
-                      Validate race condition handling
-                      - Validate that already queued items won't be scrubbed again
-        * Scenario 9: 2 vPools, 20 vDisks, 2 scrub roles
-                      Validate race condition handling
-                      - Validate that proxies will be re-used
-                      - Validate that proxies won't be removed when still in use
-        """
-        ##############
-        # Scenario 1 #
-        ##############
-        self._clean_scrubbing_test()
-        structure = DalHelper.build_dal_structure(
-            {'vpools': [1],
-             'vdisks': [(1, 1, 1, 1)],  # (<id>, <storagedriver_id>, <vpool_id>, <mds_service_id>)
-             'mds_services': [(1, 1)],  # (<id>, <storagedriver_id>)
-             'storagerouters': [1],
-             'storagedrivers': [(1, 1, 1)]}  # (<id>, <vpool_id>, <storagerouter_id>)
-        )
-        self._prepare_scrubbing(structure)
-        vdisk = structure['vdisks'][1]
-        storagerouter = structure['storagerouters'][1]
-        vpool = structure['vpools'][1]
-        LockedClient.scrub_controller = {'possible_threads': None,
-                                         'volumes': {},
-                                         'waiter': Waiter(1)}
-        LockedClient.scrub_controller['volumes'][vdisk.volume_id] = {'success': False,
-                                                                     'scrub_work': [0]}
-
-        # Remove SCRUB partition from StorageRouter and try to scrub on it
-        expected_log = 'Scrubber unittest - Assuming StorageRouter {0} is dead. Not scrubbing there'.format(storagerouter.ip)
-        storagerouter.disks[0].partitions[0].roles = []
-        storagerouter.disks[0].partitions[0].save()
-        with self.assertRaises(ValueError) as raise_info:
-            GenericController.execute_scrub(vdisk_guids=[vdisk.guid], manual=True)
-        self.assertIn(member='No scrub locations found',
-                      container=raise_info.exception.message)
-        logs = LogHandler._logs.get('lib_generic tasks scrub', [])  # No logging should have happened yet
-        self.assertNotIn(member=expected_log,
-                         container=logs)
-
-        # Restore SCRUB partition and make sure StorageRouter is unreachable
-        storagerouter.disks[0].partitions[0].roles = [DiskPartition.ROLES.SCRUB]
-        storagerouter.disks[0].partitions[0].save()
-        storagerouter.invalidate_dynamics('partition_config')
-        SSHClient._raise_exceptions[storagerouter.ip] = {'users': ['root'], 'exception': UnableToConnectException('No route to host')}
-        with self.assertRaises(ValueError):
-            GenericController.execute_scrub(vdisk_guids=[vdisk.guid], manual=True)
-        logs = LogHandler._logs['lib_generic tasks scrub']
-        self.assertIn(member=expected_log,
-                      container=logs)
-        self.assertEqual(first=logs[expected_log],
-                         second='error')
-
-        # Now actually attempt to scrub
-        # The scrub should fail as no result has been set for the vDisks
-        SSHClient._raise_exceptions = {}
-        with self.assertRaises(Exception) as raise_info:
-            GenericController.execute_scrub(vdisk_guids=[vdisk.guid], storagerouter_guid=storagerouter.guid, manual=True)
-        # Only one stack would be deployed (one scrub location) and only one thread (one vDisk)
-        expected_log = 'Scrubber unittest - vPool {0} - StorageRouter {1} - Stack 0 - Scrubbing thread 0 - vDisk {2} - Scrubbing failed'.format(vpool.name, storagerouter.name, vdisk.name)
-        self.assertIn(member=expected_log,
-                      container=raise_info.exception.message)
-
-        # Make sure scrubbing succeeds now
-        LockedClient.scrub_controller['volumes'][vdisk.volume_id] = {'success': True,
-                                                                     'scrub_work': [0]}
-        GenericController.execute_scrub(vdisk_guids=[vdisk.guid], storagerouter_guid=storagerouter.guid, manual=True)
-        with vdisk.storagedriver_client.make_locked_client(vdisk.volume_id) as locked_client:
-            self.assertEqual(first=len(locked_client.get_scrubbing_workunits()),
-                             second=0,
-                             msg='Scrubbed vDisk {0} does not have the expected amount of scrubbing items: {1}'.format(vdisk.name, 0))
-
-        ##############
-        # Scenario 2 #
-        ##############
-        self._clean_scrubbing_test()
-        structure = DalHelper.build_dal_structure(
-            {'vpools': [1],
-             'vdisks': [(1, 1, 1, 1), (2, 1, 1, 1), (3, 1, 1, 1), (4, 1, 1, 1), (5, 1, 1, 1),
-                        (6, 1, 1, 1), (7, 1, 1, 1), (8, 1, 1, 1), (9, 1, 1, 1), (10, 1, 1, 1)],  # (<id>, <storagedriver_id>, <vpool_id>, <mds_service_id>)
-             'mds_services': [(1, 1)],  # (<id>, <storagedriver_id>)
-             'storagerouters': [1, 2],
-             'storagedrivers': [(1, 1, 1)]}  # (<id>, <vpool_id>, <storagerouter_id>)
-        )
-        self._prepare_scrubbing(structure)
-        vpools = structure['vpools']
-        vpool = structure['vpools'][1]
-        vdisks = structure['vdisks']
-        storagerouter_1 = structure['storagerouters'][1]
-        storagerouter_2 = structure['storagerouters'][2]
-        LockedClient.scrub_controller = {'possible_threads': ['execute_scrub_{0}_{1}_0'.format(vpool.guid, storagerouter_1.disks[0].partitions[0].guid),
-                                                              'execute_scrub_{0}_{1}_1'.format(vpool.guid, storagerouter_1.disks[0].partitions[0].guid)],
-                                         'volumes': {},
-                                         'waiter': Waiter(1)}
-
-        # Have 1 StorageRouter with 0 SCRUB partitions
-        storagerouter_2.disks[0].partitions[0].roles = []
-        storagerouter_2.disks[0].partitions[0].save()
-
-        # Have 0 SCRUB roles and verify error
-        storagerouter_1.disks[0].partitions[0].roles = []
-        storagerouter_1.disks[0].partitions[0].save()
-        storagerouter_1.invalidate_dynamics('partition_config')
-        with self.assertRaises(ValueError) as raise_info:
-            GenericController.execute_scrub()
-        self.assertEqual(first='No scrub locations found, cannot scrub',
-                         second=raise_info.exception.message,
-                         msg='Incorrect error message caught')
-        storagerouter_1.disks[0].partitions[0].roles = [DiskPartition.ROLES.SCRUB]
-        storagerouter_1.disks[0].partitions[0].save()
-        storagerouter_1.invalidate_dynamics('partition_config')
-
-        failed_vdisks = []
-        successful_vdisks = []
-        for vdisk_id in sorted(vdisks):
-            vdisk = vdisks[vdisk_id]
-            success = vdisk_id % 2 == 0
-            LockedClient.scrub_controller['volumes'][vdisk.volume_id] = {'success': success,
-                                                                         'scrub_work': range(vdisk_id)}
-            if success is True:
-                successful_vdisks.append(vdisk)
-            else:
-                failed_vdisks.append(vdisk)
-
-        # Execute scrubbing a 1st time
-        with self.assertRaises(Exception) as raise_info:
-            GenericController.execute_scrub()
-        vdisk_namespaces = {}  # Validate all registered items are gone
-        for vpool in vpools.values():
-            vdisk_namespaces[ScrubShared._SCRUB_VDISK_KEY.format(vpool.name)] = []
-        for vdisk_namespace, namespace_value in vdisk_namespaces.iteritems():  # All registered items of this job should be cleaned up
-            self.assertEquals(self.persistent.get(vdisk_namespace), namespace_value)
-
-        for vdisk in failed_vdisks:
-            self.assertIn(vdisk.name, raise_info.exception.message)
-
-        # Validate expected successful vDisks
-        for vdisk in successful_vdisks:
-            with vdisk.storagedriver_client.make_locked_client(vdisk.volume_id) as locked_client:
-                self.assertEqual(first=len(locked_client.get_scrubbing_workunits()),
-                                 second=0,
-                                 msg='Scrubbed vDisk {0} does still have scrubbing work left'.format(vdisk.name))
-        # Validate expected failed vDisks
-        for vdisk in failed_vdisks:
-            with vdisk.storagedriver_client.make_locked_client(vdisk.volume_id) as locked_client:
-                self.assertEqual(first=len(locked_client.get_scrubbing_workunits()),
-                                 second=int(vdisk.name),
-                                 msg='Scrubbed vDisk {0} does not have the expected amount of scrubbing items: {1}'.format(vdisk.name, int(vdisk.name)))
-
-        # Execute scrubbing again
-        for vdisk_id in sorted(vdisks):
-            vdisk = vdisks[vdisk_id]
-            LockedClient.scrub_controller['volumes'][vdisk.volume_id]['success'] = True
-        GenericController.execute_scrub()
-        vdisk_namespaces = {}  # Validate all registered items are gone
-        for vpool in vpools.values():
-            vdisk_namespaces[ScrubShared._SCRUB_VDISK_KEY.format(vpool.name)] = []
-        for vdisk_namespace, namespace_value in vdisk_namespaces.iteritems():  # All registered items of this job should be cleaned up
-            self.assertEquals(self.persistent.get(vdisk_namespace), namespace_value)
-
-        for vdisk in vdisks.values():
-            with vdisk.storagedriver_client.make_locked_client(vdisk.volume_id) as locked_client:
-                self.assertEqual(first=len(locked_client.get_scrubbing_workunits()),
-                                 second=0,
-                                 msg='Scrubbed vDisk {0} does still have scrubbing work left after scrubbing a 2nd time'.format(vdisk.name))
-
-        ##############
-        # Scenario 3 #
-        ##############
-        self._clean_scrubbing_test()
-        structure = DalHelper.build_dal_structure(
-            {'vpools': [1, 2],  # vPool 2 has no vDisks attached to it
-             'vdisks': [(i, 1, 1, 1) for i in xrange(1, 12)],  # (<id>, <storagedriver_id>, <vpool_id>, <mds_service_id>)
-             'mds_services': [(1, 1)],  # (<id>, <storagedriver_id>)
-             'storagerouters': [1, 2, 3, 4],
-             'storagedrivers': [(1, 1, 1)]}  # (<id>, <vpool_id>, <storagerouter_id>)
-        )
-        self._prepare_scrubbing(structure)
-        vpool = structure['vpools'][1]
-        vpools = structure['vpools']
-        vdisks = structure['vdisks']
-        storagerouter_1 = structure['storagerouters'][1]
-        # Have 1 volume as a template, scrubbing should not be triggered on it
-        vdisk_t = structure['vdisks'][11]
-        vdisk_t.storagedriver_client.set_volume_as_template(volume_id=vdisk_t.volume_id)
-
-        # Have 1 StorageRouter with multiple SCRUB partitions
-        partition = DiskPartition()
-        partition.offset = 0
-        partition.size = storagerouter_1.disks[0].size
-        partition.aliases = ['/dev/uda-2']
-        partition.state = 'OK'
-        partition.mountpoint = '/tmp/unittest/sr_1/disk_1/partition_2'
-        partition.disk = storagerouter_1.disks[0]
-        partition.roles = [DiskPartition.ROLES.SCRUB]
-        partition.save()
-
-        thread_names = []
-        for storagerouter in structure['storagerouters'].values():
-            for partition in [p for disk in storagerouter.disks for p in disk.partitions]:
-                for index in range(2):
-                    thread_names.append('execute_scrub_{0}_{1}_{2}'.format(vpool.guid, partition.guid, index))
-        LockedClient.scrub_controller = {'possible_threads': thread_names,
-                                         'volumes': {},
-                                         'waiter': Waiter(len(thread_names))}
-        LockedClient.thread_names = thread_names[:]
-        for vdisk_id in sorted(vdisks):
-            vdisk = vdisks[vdisk_id]
-            LockedClient.scrub_controller['volumes'][vdisk.volume_id] = {'success': True,
-                                                                         'scrub_work': range(vdisk_id)}
-        GenericController.execute_scrub()
-        vdisk_namespaces = {}  # Validate all registered items are gone
-        for vpool in vpools.values():
-            vdisk_namespaces[ScrubShared._SCRUB_VDISK_KEY.format(vpool.name)] = []
-        for vdisk_namespace, namespace_value in vdisk_namespaces.iteritems():  # All registered items of this job should be cleaned up
-            self.assertEquals(self.persistent.get(vdisk_namespace), namespace_value)
-
-        # Verify all threads have been 'consumed'
-        self.assertEqual(first=len(LockedClient.thread_names),
-                         second=0)
-        self.assertIn(member='Scrubber unittest - vPool {0} - vDisk {1} with guid {2} is a template, not scrubbing'.format(vdisk_t.vpool.name, vdisk_t.name, vdisk_t.guid),
-                      container=LogHandler._logs['lib_generic tasks scrub'])
-
-        ##############
-        # Scenario 4 #
-        ##############
-        self._clean_scrubbing_test()
-        structure = DalHelper.build_dal_structure(
-            {'vpools': [1, 2, 3],
-             'vdisks': [(1, 1, 1, 1), (2, 1, 1, 1), (3, 1, 1, 1), (4, 1, 1, 1), (5, 1, 1, 1),
-                        (6, 2, 2, 2), (7, 2, 2, 2), (8, 2, 2, 2), (9, 2, 2, 2), (10, 2, 2, 2),
-                        (11, 3, 3, 3), (12, 3, 3, 3), (13, 3, 3, 3), (14, 3, 3, 3), (15, 3, 3, 3)],  # (<id>, <storagedriver_id>, <vpool_id>, <mds_service_id>)
-             'mds_services': [(1, 1), (2, 2), (3, 3)],  # (<id>, <storagedriver_id>)
-             'storagerouters': [1, 2, 3, 4, 5],
-             'storagedrivers': [(1, 1, 1), (2, 2, 1), (3, 3, 1)]}  # (<id>, <vpool_id>, <storagerouter_id>)
-        )
-        self._prepare_scrubbing(structure)
-        vpools = structure['vpools']
-        vdisks = structure['vdisks']
-        storagerouters = structure['storagerouters']
-        # Amount of actual threads calculation:
-        #   - Threads per VPool * vPools * 2 threads per StorageRouter
-        #   - Threads per vPool is 2 when 3 vPools and 5 StorageRouters
-        #   - Amount of threads that will be created: 2 * 3 * 2 = 12
-        # Amount of possible threads calculation:
-        #   - vPools * StorageRouters * 2 threads per StorageRouter
-        #   - Amount of possible threads to be created: 3 * 5 * 2 = 30
-        thread_names = []
-        for vpool in vpools.values():
-            for storagerouter in storagerouters.values():
-                for partition in storagerouter.disks[0].partitions:
-                    for index in range(2):
-                        thread_names.append('execute_scrub_{0}_{1}_{2}'.format(vpool.guid, partition.guid, index))
-        LockedClient.scrub_controller = {'possible_threads': thread_names,
-                                         'volumes': {},
-                                         'waiter': Waiter(12)}
-        LockedClient.thread_names = thread_names[:]
-        for vdisk_id in sorted(vdisks):
-            vdisk = vdisks[vdisk_id]
-            LockedClient.scrub_controller['volumes'][vdisk.volume_id] = {'success': True,
-                                                                         'scrub_work': range(vdisk_id)}
-        GenericController.execute_scrub()
-        vdisk_namespaces = {}  # Validate all registered items are gone
-        for vpool in vpools.values():
-            vdisk_namespaces[ScrubShared._SCRUB_VDISK_KEY.format(vpool.name)] = []
-        for vdisk_namespace, namespace_value in vdisk_namespaces.iteritems():  # All registered items of this job should be cleaned up
-            self.assertEquals(self.persistent.get(vdisk_namespace), namespace_value)
-
-        self.assertEqual(first=len(LockedClient.thread_names),
-                         second=18,  # 30 possible thread_names - 12 which should be created and consumed
-                         msg='Not all threads have been used in the process')
-
-        # Of the 18 threads which have been created and consumed, 6 should have been created for each vPool
-        for vpool in vpools.values():
-            threads_left = [thread_name for thread_name in LockedClient.thread_names if vpool.guid in thread_name]
-            self.assertEqual(first=len(threads_left),
-                             second=6,
-                             msg='Unexpected amount of threads left for vPool {0}'.format(vpool.name))
-
-        ##############
-        # Scenario 5 #
-        ##############
-        self._clean_scrubbing_test()
-        structure = DalHelper.build_dal_structure(
-            {'vpools': [1, 2],
-             'vdisks': [(1, 1, 1, 1), (2, 2, 1, 2), (3, 3, 2, 3), (4, 4, 2, 4),
-                        (5, 1, 1, 1), (6, 2, 1, 2), (7, 3, 2, 3), (8, 4, 2, 4)],  # (<id>, <storagedriver_id>, <vpool_id>, <mds_service_id>)
-             'mds_services': [(1, 1), (2, 2), (3, 3), (4, 4)],  # (<id>, <storagedriver_id>)
-             'storagerouters': [1, 2],
-             'storagedrivers': [(1, 1, 1), (2, 1, 2), (3, 2, 1), (4, 2, 2)]}  # (<id>, <vpool_id>, <storagerouter_id>)
-        )
-        self._prepare_scrubbing(structure)
-        vpools = structure['vpools']
-        vdisks = structure['vdisks']
-        storagerouters = structure['storagerouters']
-        # Amount of actual threads calculation:
-        #   - Threads per VPool * vPools * 2 threads per StorageRouter
-        #   - Threads per vPool is 2 when 2 vPools and 2 StorageRouters
-        #   - Amount of threads that will be created: 2 * 2 * 2 = 8
-        # Amount of possible threads calculation:
-        #   - vPools * StorageRouters * 2 threads per StorageRouter
-        #   - Amount of possible threads to be created: 2 * 2 * 2 = 8
-        thread_names = []
-        for vpool in vpools.values():
-            for storagerouter in storagerouters.values():
-                for partition in storagerouter.disks[0].partitions:
-                    for index in range(2):
-                        thread_names.append('execute_scrub_{0}_{1}_{2}'.format(vpool.guid, partition.guid, index))
-        LockedClient.scrub_controller = {'possible_threads': thread_names,
-                                         'volumes': {},
-                                         'waiter': Waiter(len(thread_names))}
-
-        # Scrub all volumes
-        for vdisk_id, vdisk in vdisks.iteritems():
-            LockedClient.scrub_controller['volumes'][vdisk.volume_id] = {'success': True,
-                                                                         'scrub_work': range(vdisk_id)}
-        GenericController.execute_scrub()
-        vdisk_namespaces = {}  # Validate all registered items are gone
-        for vpool in vpools.values():
-            vdisk_namespaces[ScrubShared._SCRUB_VDISK_KEY.format(vpool.name)] = []
-        for vdisk_namespace, namespace_value in vdisk_namespaces.iteritems():  # All registered items of this job should be cleaned up
-            self.assertEquals(self.persistent.get(vdisk_namespace), namespace_value)
-
-        for vdisk in vdisks.values():
-            self.assertListEqual(list1=LockedClient.scrub_controller['volumes'][vdisk.volume_id]['scrub_work'],
-                                 list2=[])
-
-        # Scrub all volumes of vPool1
-        for vdisk_id, vdisk in vdisks.iteritems():
-            LockedClient.scrub_controller['volumes'][vdisk.volume_id]['scrub_work'] = range(vdisk_id)
-        GenericController.execute_scrub(vpool_guids=[vpools[1].guid], manual=True)
-        for vdisk_id, vdisk in vdisks.iteritems():
-            if vdisk.vpool == vpools[1]:
-                expected_work = []
-            else:
-                expected_work = range(vdisk_id)
-            self.assertListEqual(list1=LockedClient.scrub_controller['volumes'][vdisk.volume_id]['scrub_work'],
-                                 list2=expected_work)
-
-        # Scrub a specific vDisk
-        for vdisk_id, vdisk in vdisks.iteritems():
-            LockedClient.scrub_controller['volumes'][vdisk.volume_id]['scrub_work'] = range(vdisk_id)
-        GenericController.execute_scrub(vdisk_guids=[vdisks[2].guid], manual=True)
-        for vdisk_id, vdisk in vdisks.iteritems():
-            if vdisk == vdisks[2]:
-                expected_work = []
-            else:
-                expected_work = range(vdisk_id)
-            self.assertListEqual(list1=LockedClient.scrub_controller['volumes'][vdisk.volume_id]['scrub_work'],
-                                 list2=expected_work,
-                                 msg='Not all expected work items were applied for vDisk {0}'.format(vdisk.name))
-
-        # Scrub a combination of a vPool and a vDisk
-        for vdisk_id, vdisk in vdisks.iteritems():
-            LockedClient.scrub_controller['volumes'][vdisk.volume_id]['scrub_work'] = range(vdisk_id)
-        GenericController.execute_scrub(vpool_guids=[vpools[2].guid], vdisk_guids=[vdisks[2].guid], manual=True)
-        for vdisk_id, vdisk in vdisks.iteritems():
-            if vdisk == vdisks[2] or vdisk.vpool == vpools[2]:
-                expected_work = []
-            else:
-                expected_work = range(vdisk_id)
-            self.assertListEqual(list1=LockedClient.scrub_controller['volumes'][vdisk.volume_id]['scrub_work'],
-                                 list2=expected_work)
-
-        # Scrub all volumes on specific StorageRouter
-        LogHandler._logs = {}
-        for vdisk_id, vdisk in vdisks.iteritems():
-            LockedClient.scrub_controller['volumes'][vdisk.volume_id]['scrub_work'] = range(vdisk_id)
-        GenericController.execute_scrub(storagerouter_guid=storagerouters[2].guid)
-        for vdisk_id, vdisk in vdisks.iteritems():
-            self.assertListEqual(list1=LockedClient.scrub_controller['volumes'][vdisk.volume_id]['scrub_work'],
-                                 list2=[],
-                                 msg='Not all scrub work was applied for vDisk {0}'.format(vdisk.name))
-        logs = LogHandler._logs['lib_generic tasks scrub']
-        for log in logs:
-            self.assertNotRegexpMatches(text=log,
-                                        unexpected_regexp='.*Scrubber unittest - vPool [{0}|{1}] - StorageRouter {2} - .*'.format(vpools[1].name, vpools[2].name, storagerouters[1].name))
-
-        ##############
-        # Scenario 6 #
-        ##############
-        self._clean_scrubbing_test()
-        structure = DalHelper.build_dal_structure(
-            {'vpools': [1, 2],
-             'vdisks': [(1, 1, 1, 1), (2, 1, 1, 1), (3, 1, 1, 1), (4, 1, 1, 1), (5, 1, 1, 1),
-                        (6, 2, 1, 2), (7, 2, 1, 2), (8, 2, 1, 2), (9, 2, 1, 2), (10, 2, 1, 2),
-                        (11, 3, 2, 3), (12, 3, 2, 3), (13, 3, 2, 3), (14, 3, 2, 3), (15, 3, 2, 3),
-                        (16, 4, 2, 4), (17, 4, 2, 4), (18, 4, 2, 4), (19, 4, 2, 4), (20, 4, 2, 4)],  # (<id>, <storagedriver_id>, <vpool_id>, <mds_service_id>)
-             'mds_services': [(1, 1), (2, 2), (3, 3), (4, 4)],  # (<id>, <storagedriver_id>)
-             'storagerouters': [1, 2],
-             'storagedrivers': [(1, 1, 1), (2, 1, 2), (3, 2, 1), (4, 2, 2)]}  # (<id>, <vpool_id>, <storagerouter_id>)
-        )
-        self._prepare_scrubbing(structure)
-        vpools = structure['vpools']
-        vdisks = structure['vdisks']
-        storagerouters = structure['storagerouters']
-        # Set amount of stack threads for SR1 to 5 and leave for SR2 to default (2)
-        sr_1_threads = 5
-        sr_2_threads = 2
-        Configuration.set(key='/ovs/framework/hosts/{0}/config|scrub_stack_threads'.format(storagerouters[1].machine_id), value=sr_1_threads)
-        # Amount of actual threads calculation:
-        #   - Threads per VPool * vPools * <scrub_stack_threads> threads per StorageRouter
-        #   - Threads per vPool is 2 when 2 vPools and 2 StorageRouters
-        #   - Amount of threads that will be created:  2 * 2 * (2 + 5) / 2 = 14
-        #       - For StorageRouter 1: 10
-        #       - For StorageRouter 2: 4
-        # Amount of possible threads calculation:
-        #   - vPools * StorageRouters * <scrub_stack_threads> threads per StorageRouter
-        #   - Amount of possible threads to be created: 2 * 2 * (2 + 5) / 2 = 14
-        thread_names = []
-        for vpool in vpools.values():
-            for storagerouter in storagerouters.values():
-                stack_threads = sr_1_threads if storagerouter == storagerouters[1] else sr_2_threads
-                for partition in storagerouter.disks[0].partitions:
-                    for index in range(stack_threads):
-                        thread_names.append('execute_scrub_{0}_{1}_{2}'.format(vpool.guid, partition.guid, index))
-        LockedClient.scrub_controller = {'possible_threads': thread_names,
-                                         'volumes': {},
-                                         'waiter': Waiter(14)}
-        LockedClient.thread_names = thread_names[:]
-        for vdisk_id in sorted(vdisks):
-            vdisk = vdisks[vdisk_id]
-            LockedClient.scrub_controller['volumes'][vdisk.volume_id] = {'success': True,
-                                                                         'scrub_work': range(vdisk_id)}
-        LogHandler._logs = {}
-        GenericController.execute_scrub()
-        # Verify all threads have been 'consumed'
-        self.assertEqual(first=len(LockedClient.thread_names),
-                         second=0)
-        vdisk_namespaces = {}  # Validate all registered items are gone
-        for vpool in vpools.values():
-            vdisk_namespaces[ScrubShared._SCRUB_VDISK_KEY.format(vpool.name)] = []
-        for vdisk_namespace, namespace_value in vdisk_namespaces.iteritems():  # All registered items of this job should be cleaned up
-            self.assertEquals(self.persistent.get(vdisk_namespace), namespace_value)
-        counter = 0
-        for log in LogHandler._logs['lib_generic tasks scrub']:
-            if 'threads for proxy service' in log:
-                match = re.match('^Scrubber unittest - vPool ([1|2]) - StorageRouter ([1|2]) - Stack ([0|1]) - .*ovs-albaproxy.*_scrub', log)
-                self.assertIsNotNone(match)
-                if match.groups()[1] == storagerouters[1].name:
-                    expected_threads = 5
-                else:
-                    expected_threads = 2
-                self.assertIn(member='Spawning {0} threads for proxy'.format(expected_threads),
-                              container=log)
-                counter += 1
-        self.assertEqual(first=4,  # Log entry for each combination of 2 vPools and 2 StorageRouters
-                         second=counter)
-
-        ##############
-        # Scenario 7 #
-        ##############
-        def _raise_exception(message):
-            raise RuntimeError(message)
-
-        self._clean_scrubbing_test()
-        structure = DalHelper.build_dal_structure(
-            {'vpools': [1, 2],
-             'vdisks': [(1, 1, 1, 1), (2, 1, 1, 1), (3, 1, 1, 1), (4, 1, 1, 1), (5, 1, 1, 1),
-                        (6, 2, 1, 2), (7, 2, 1, 2), (8, 2, 1, 2), (9, 2, 1, 2), (10, 2, 1, 2),
-                        (11, 3, 2, 3), (12, 3, 2, 3), (13, 3, 2, 3), (14, 3, 2, 3), (15, 3, 2, 3),
-                        (16, 4, 2, 4), (17, 4, 2, 4), (18, 4, 2, 4), (19, 4, 2, 4), (20, 4, 2, 4)],
-             # (<id>, <storagedriver_id>, <vpool_id>, <mds_service_id>)
-             'mds_services': [(1, 1), (2, 2), (3, 3), (4, 4)],  # (<id>, <storagedriver_id>)
-             'storagerouters': [1, 2],
-             'storagedrivers': [(1, 1, 1), (2, 1, 2), (3, 2, 1), (4, 2, 2)]}  # (<id>, <vpool_id>, <storagerouter_id>)
-        )
-        self._prepare_scrubbing(structure)
-        vpools = structure['vpools']
-        vdisks = structure['vdisks']
-        storagerouters = structure['storagerouters']
-        # Amount of actual threads calculation:
-        #   - Threads per VPool * vPools * 2 threads per StorageRouter
-        #   - Threads per vPool is 2 when 2 vPools and 2 StorageRouters
-        #   - Amount of threads that will be created: 2 * 2 * 2 = 8
-        # Amount of possible threads calculation:
-        #   - vPools * StorageRouters * 2 threads per StorageRouter
-        #   - Amount of possible threads to be created: 2 * 2 * 2 = 8
-        thread_names = []
-        proxy_names = []
-        vdisk_namespaces = {}
-        for vpool in vpools.values():
-            vdisk_namespaces[ScrubShared._SCRUB_VDISK_KEY.format(vpool.name)] = []
-            for storagerouter in storagerouters.values():
-                for partition in storagerouter.disks[0].partitions:
-                    proxy_name = 'ovs-albaproxy_{0}_{1}_{2}_scrub'.format(vpool.name, storagerouter.name, partition.guid)
-                    proxy_names.append(proxy_name)
-                    for index in range(2):
-                        thread_names.append('execute_scrub_{0}_{1}_{2}'.format(vpool.guid, partition.guid, index))
-        LockedClient.scrub_controller = {'possible_threads': thread_names,
-                                         'volumes': {},
-                                         'waiter': Waiter(len(thread_names))}
-
-        # Scrub all volumes
-        for vdisk_id, vdisk in vdisks.iteritems():
-            LockedClient.scrub_controller['volumes'][vdisk.volume_id] = {'success': True,
-                                                                         'scrub_work': range(vdisk_id)}
-        # Insert some hooks to create some failure cases
-        hooks = {'post_proxy_deployment': lambda x: _raise_exception('Simulated proxy deployment failure')}
-        ScrubShared._test_hooks.update(hooks)
-        # No scrubbing should have taken place
-        with self.assertRaises(Exception):
-            GenericController.execute_scrub()
-        # Check if service has been removed properly
-        service_manager = ServiceFactory.get_manager()
-        for storagerouter in storagerouters.values():
-            client = SSHClient(storagerouter, 'root')
-            for proxy_name in proxy_names:
-                self.assertFalse(service_manager.has_service(proxy_name, client), 'StorageRouter {0} still has proxy {1}'.format(storagerouter.name, proxy_name))
-        for vdisk_namespace, namespace_value in vdisk_namespaces.iteritems():  # All registered items of this job should be cleaned up
-            self.assertEquals(self.persistent.get(vdisk_namespace), namespace_value)
-        ##############
-        # Scenario 8 #
-        ##############
-        # Test queue work detection
-        self._clean_scrubbing_test()
-        structure = DalHelper.build_dal_structure(
-            {'vpools': [1, 2],
-             # (<id>, <storagedriver_id>, <vpool_id>, <mds_service_id>)
-             'vdisks': [(1, 1, 1, 1), (2, 1, 1, 1), (3, 1, 1, 1), (4, 1, 1, 1), (5, 1, 1, 1),
-                        (6, 2, 1, 2), (7, 2, 1, 2), (8, 2, 1, 2), (9, 2, 1, 2), (10, 2, 1, 2),
-                        (11, 3, 2, 3), (12, 3, 2, 3), (13, 3, 2, 3), (14, 3, 2, 3), (15, 3, 2, 3),
-                        (16, 4, 2, 4), (17, 4, 2, 4), (18, 4, 2, 4), (19, 4, 2, 4), (20, 4, 2, 4)],
-             'mds_services': [(1, 1), (2, 2), (3, 3), (4, 4)],  # (<id>, <storagedriver_id>)
-             'storagerouters': [1, 2],
-             'storagedrivers': [(1, 1, 1), (2, 1, 2), (3, 2, 1), (4, 2, 2)]}  # (<id>, <vpool_id>, <storagerouter_id>)
-        )
-        self._prepare_scrubbing(structure)
-        vpools = structure['vpools']
-        vdisks = structure['vdisks']
-        storagerouters = structure['storagerouters']
-        thread_names = []
-        proxy_names = []
-        vdisk_namespaces = {}
-        # Thread calculation is the same as Scenario 7
-        for vpool in vpools.values():
-            vdisk_namespaces[ScrubShared._SCRUB_VDISK_KEY.format(vpool.name)] = []
-            for storagerouter in storagerouters.values():
-                for partition in storagerouter.disks[0].partitions:
-                    proxy_name = 'ovs-albaproxy_{0}_{1}_{2}_scrub'.format(vpool.name, storagerouter.name, partition.guid)
-                    proxy_names.append(proxy_name)
-                    for index in range(2):
-                        thread_names.append('execute_scrub_{0}_{1}_{2}'.format(vpool.guid, partition.guid, index))
-        LockedClient.scrub_controller = {'possible_threads': thread_names,
-                                         'volumes': {},
-                                         'waiter': Waiter(len(thread_names))}
-
-        # Scrub all volumes
-        for vdisk_id, vdisk in vdisks.iteritems():
-            LockedClient.scrub_controller['volumes'][vdisk.volume_id] = {'success': True,
-                                                                         'scrub_work': range(vdisk_id)}
-        # Test already queued items
-        local_sr = System.get_my_storagerouter()  # Executor of the scrubbing
-        pid = None
-        start_time = None
-        other_sr = None
-        for ip, service_info in SystemdMock.services.iteritems():
-            if ip == local_sr.ip:
-                continue
-            other_sr = StorageRouterList.get_by_ip(ip)
-            worker_info = service_info.get('ovs-workers', {})
-            pid = worker_info.get('pid')
-            start_time = worker_info.get('start_time')
-            break
-        self.assertNotEqual(pid, None, 'A PID should be found for a different StorageRouter\'s ovs-worker process')
-        self.assertNotEqual(start_time, None, 'A start time should be found for a different StorageRouter\'s ovs-worker process')
-        # Register vdisks beforehand
-        for index, vdisk in enumerate(vdisks.values()):
-            if index == 2:
-                break
-            for vdisk_namespace, namespace_value in vdisk_namespaces.iteritems():
-                namespace_value.append({'vdisk_guid': vdisk.guid, 'worker_pid': pid, 'worker_start': start_time, 'storagerouter_guid': other_sr.guid})
-                self.persistent.set(vdisk_namespace, namespace_value)
-        GenericController.execute_scrub()
-        logs = LogHandler._logs['lib_generic tasks scrub']
-        for registered_items in vdisk_namespaces.values():
-            for registered_item in registered_items:
-                vdisk = VDisk(registered_item['vdisk_guid'])
-                log = 'Scrubber unittest - vPool {0} - vDisk {1} with guid {2} has already been registered to get scrubbed, not queueing again'.format(vdisk.vpool.name, vdisk.name, vdisk.guid)
-                self.assertIn(log, logs)
-
-        ##############
-        # Scenario 9 #
-        ##############
-        stack_workers = []  # Copied stack workers by hooking
-
-        def _wait_for_event(event):
-            """
-            Wait forever until the event is set
-            :param event: Event object
-            :type event: Threading.Event
-            :return: None
-            :rtype: NoneType
-            """
-            event.wait()
-
-        def _inject_race(self):
-            """
-            Injects a race condition. The stack worker threads will wait for another the event to clear.
-            A stackworker is taken to duplicate it and register its proxy usage
-            :param self: Scrubber instance
-            :return: None
-            :rtype: NoneType
-            """
-            for model_stack_worker in self.stack_workers:
-                stack_worker = StackWorker(queue=model_stack_worker.queue,
-                                           vpool=model_stack_worker.vpool,
-                                           scrub_info=model_stack_worker.scrub_info,
-                                           error_messages=self.error_messages,
-                                           worker_contexts=self.worker_contexts,
-                                           stack_work_handler=model_stack_worker.stack_work_handler,
-                                           job_id='Another job',
-                                           stacks_to_spawn=100,
-                                           stack_number=100)
-                stack_workers.append(stack_worker)
-                stack_worker._test_hooks = {}  # Clear hooks
-                stack_worker._deploy_proxy()
-            proxy_event.set()  # Allow others to deploy proxies now
-
-        # Test proxy re-use
-        # Test is use proxies won't be removed
-        self._clean_scrubbing_test()
-        structure = DalHelper.build_dal_structure(
-            {'vpools': [1, 2],
-             # (<id>, <storagedriver_id>, <vpool_id>, <mds_service_id>)
-             'vdisks': [(1, 1, 1, 1), (2, 1, 1, 1), (3, 1, 1, 1), (4, 1, 1, 1), (5, 1, 1, 1),
-                        (6, 2, 1, 2), (7, 2, 1, 2), (8, 2, 1, 2), (9, 2, 1, 2), (10, 2, 1, 2),
-                        (11, 3, 2, 3), (12, 3, 2, 3), (13, 3, 2, 3), (14, 3, 2, 3), (15, 3, 2, 3),
-                        (16, 4, 2, 4), (17, 4, 2, 4), (18, 4, 2, 4), (19, 4, 2, 4), (20, 4, 2, 4)],
-             'mds_services': [(1, 1), (2, 2), (3, 3), (4, 4)],  # (<id>, <storagedriver_id>)
-             'storagerouters': [1, 2],
-             'storagedrivers': [(1, 1, 1), (2, 1, 2), (3, 2, 1), (4, 2, 2)]}  # (<id>, <vpool_id>, <storagerouter_id>)
-        )
-        self._prepare_scrubbing(structure)
-        vpools = structure['vpools']
-        vdisks = structure['vdisks']
-        storagerouters = structure['storagerouters']
-        thread_names = []
-        proxy_names = []
-        # Thread calculation is the same as Scenario 7
-        for vpool in vpools.values():
-            for storagerouter in storagerouters.values():
-                for partition in storagerouter.disks[0].partitions:
-                    proxy_names.append('ovs-albaproxy_{0}_{1}_{2}_scrub'.format(vpool.name, storagerouter.name, partition.guid))
-                    for index in range(2):
-                        thread_names.append('execute_scrub_{0}_{1}_{2}'.format(vpool.guid, partition.guid, index))
-        LockedClient.scrub_controller = {'possible_threads': thread_names,
-                                         'volumes': {},
-                                         'waiter': Waiter(len(thread_names))}
-
-        # Scrub all volumes
-        for vdisk_id, vdisk in vdisks.iteritems():
-            LockedClient.scrub_controller['volumes'][vdisk.volume_id] = {'success': True,
-                                                                         'scrub_work': range(vdisk_id)}
-        proxy_event = Event()
-        hooks = {'pre_proxy_deployment': lambda x: _wait_for_event(proxy_event),  # This will make the scrubber stacks to wait forever before deploying proxies
-                 'post_stack_worker_deployment': lambda x: _inject_race(x)}
-        ScrubShared._test_hooks.update(hooks)
-        GenericController.execute_scrub()
-        logs = LogHandler._logs['lib_generic tasks scrub']
-        re_use_logs = ['Re-using existing proxy service {0}'.format(proxy_name) for proxy_name in proxy_names]
-        remove_logs = ['Cannot remove service {0} as it is still in use by others'.format(proxy_name) for proxy_name in proxy_names]
-        self.assertEquals(len([log for log in logs if log.endswith(tuple(re_use_logs))]), 4, 'All deployed stacks should be re-using the proxies')
-        self.assertEquals(len([log for log in logs if log.endswith(tuple(remove_logs))]), 4, 'All deployed stacks should be not be touching the proxies')
-        service_manager = ServiceFactory.get_manager()
-        for stack_worker in stack_workers:
-            client = SSHClient(stack_worker.storagerouter, 'root')
-            self.assertTrue(service_manager.get_service_status(name=stack_worker.alba_proxy_service, client=client) == 'active', 'The proxy should be still running')
-        for stack_worker in stack_workers:
-            stack_worker._remove_proxy()
-        for stack_worker in stack_workers:
-            client = SSHClient(stack_worker.storagerouter, 'root')
-            self.assertFalse(stack_worker._service_manager.has_service(name=stack_worker.alba_proxy_service, client=client), 'The proxy should be removed')
 
     def test_arakoon_collapse(self):
         """
@@ -1278,48 +570,3 @@ class Generic(unittest.TestCase):
     @staticmethod
     def _from_timestamp(timestamp):
         return datetime.datetime.fromtimestamp(timestamp).strftime('%Y-%m-%d %H:%M')
-
-    @classmethod
-    def _prepare_scrubbing(cls, dal_structure):
-        """
-        Prepare everything related to scrubbing
-        - Setup worker services
-        - Setup System.get_my_storagerouter
-        :param dal_structure: The built DAL structure for the case
-        :return: None
-        :rtype: NoneType
-        """
-        storagerouter = dal_structure['storagerouters'][1]
-        # New scrubbing code changes requires local Storagerouter to be available
-        System._machine_id['none'] = System._machine_id[storagerouter.ip]
-        # Setup worker information
-        cls._setup_worker_service(dal_structure['storagerouters'].values())
-
-    @staticmethod
-    def _setup_worker_service(storagerouters):
-        """
-        Sets mocked ovs-worker service
-        :param storagerouters: StorageRouter to setup workers for
-        :type storagerouters: list
-        :return: None
-        :rtype: NoneType
-        """
-        service_name = 'ovs-workers'
-        service_manager = ServiceFactory.get_manager()
-        for storagerouter in storagerouters:
-            client = SSHClient(storagerouter, 'root')
-            service_manager.add_service(service_name, client)
-            service_manager.start_service(service_name, client)
-
-    def _clean_scrubbing_test(self):
-        """
-        Clean all affected items for scrubbing
-        :return: None
-        :rtype: NoneType
-        """
-        self.volatile._clean()
-        self.persistent._clean()
-        SystemdMock._clean()
-        Configuration._unittest_data = {}
-        ScrubShared._test_hooks = {}
-        LogHandler._logs = {}
